Fixes for subtypes
Change key method (used to find CRD that backs the schema) to use
baseType of the schema when it differs from the schema ID. This
will allow subtypes to use their base type's CRD store.

When creating CRD stores, do not create one for a schema if it is
a baseType.