Use request.PostForm instead of request.Form when
PostForm will ignore URL query parameters and will only look at the
body. This is much more useful for actions, which have a query param
that is the action name, but the a json body.  WIthout this change,
the only "field" found when parsing an action body is the name of the
action.