Log panics as error
The defaul http server logs panics to the server's log handler.
That goes to debug level and thus panics are not seen when running
rancher at a normal log level.

This fixes the problem by catching the panic at the top level norman
http handler and logging them as errors there.