Don't update type=password fields with empty values
When some CRD is updated from UI, the secret fields previously provided
and stored are hidden and not passed down the request. The call is made
with replace?true, so this ends up replacing secret fields with empty values.
This commit checks whether a field of type=password (secret) is being
replaced with empty value due to an update, and doesn't update the field.