Add refresh function to Store
Added refresh function to Stores. The intention of this function
is to optionally provide logic in a store for sending an object
through the websocket consumed by the UI. Prior, formatters may
have cause an object to be different but would not be triggered
without refreshing the page. Now, views can be updated without
refreshing.